Move resource from redux to local state
