+++ src/components/resources/helpers/resourceFilterReducer.ts
@@ -15,7 +15,6 @@ const UPDATE_FILTER_ASSET = 'UPDATE_FILTER_ASSET';
 ////////////////////////////////////////////////////////////////////////////////
 // Type definitions
 ////////////////////////////////////////////////////////////////////////////////
-
 type ResourceFilterAssetState = {
     filteredResources: Resource[],
     typeAssets: Asset[],
